Fix slotchange event tests. Move slot content manip to the distribute module.
